@@ -31,8 +31,8 @@
     "@babel/core": "^7.12.3",
     "@babel/plugin-transform-runtime": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
-    "@vue/babel-helper-vue-jsx-merge-props": "^1.2.1",
-    "@vue/babel-preset-jsx": "^1.2.4",
+    "@vue/babel-plugin-jsx": "^1.0.0-rc.3",
+    "@vue/compiler-sfc": "^3.0.2",
     "autoprefixer": "^10.0.1",
     "babel-loader": "^8.1.0",
     "chalk": "^4.1.0",
@@ -64,10 +64,9 @@
     "semver": "^7.3.2",
     "shelljs": "^0.8.4",
     "url-loader": "^4.1.0",
-    "vue": "^2.6.11",
-    "vue-loader": "^15.9.4",
+    "vue": "^3.0.2",
+    "vue-loader": "^16.0.0-rc.1",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.11",
     "webpack": "^5.3.0",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-dev-middleware": "^3.7.2",
