@@ -2,56 +2,56 @@
   <div>
     <div class="block">
       <h3>JSON:</h3>
-      <textarea v-model="val" />
+      <textarea v-model="state.val" />
 
       <h3>Options:</h3>
       <div class="options">
         <div>
           <label>selectableType</label>
-          <select v-model="selectableType">
+          <select v-model="state.selectableType">
             <option>single</option>
             <option>multiple</option>
           </select>
         </div>
         <div>
           <label>showSelectController</label>
-          <input v-model="showSelectController" type="checkbox" />
+          <input v-model="state.showSelectController" type="checkbox" />
         </div>
         <div>
           <label>selectOnClickNode</label>
-          <input v-model="selectOnClickNode" type="checkbox" />
+          <input v-model="state.selectOnClickNode" type="checkbox" />
         </div>
         <div>
           <label>path</label>
-          <input v-model="path" type="text" />
+          <input v-model="state.path" type="text" />
         </div>
         <div>
           <label>showLength</label>
-          <input v-model="showLength" type="checkbox" />
+          <input v-model="state.showLength" type="checkbox" />
         </div>
         <div>
           <label>showLine</label>
-          <input v-model="showLine" type="checkbox" />
+          <input v-model="state.showLine" type="checkbox" />
         </div>
         <div>
           <label>showDoubleQuotes</label>
-          <input v-model="showDoubleQuotes" type="checkbox" />
+          <input v-model="state.showDoubleQuotes" type="checkbox" />
         </div>
         <div>
           <label>highlightMouseoverNode</label>
-          <input v-model="highlightMouseoverNode" type="checkbox" />
+          <input v-model="state.highlightMouseoverNode" type="checkbox" />
         </div>
         <div>
           <label>highlightSelectedNode</label>
-          <input v-model="highlightSelectedNode" type="checkbox" />
+          <input v-model="state.highlightSelectedNode" type="checkbox" />
         </div>
         <div>
           <label>collapsedOnClickBrackets</label>
-          <input v-model="collapsedOnClickBrackets" type="checkbox" />
+          <input v-model="state.collapsedOnClickBrackets" type="checkbox" />
         </div>
         <div>
           <label>deep</label>
-          <select v-model="deep">
+          <select v-model="state.deep">
             <option :value="2">
               2
             </option>
@@ -65,38 +65,38 @@
         </div>
         <div>
           <label>use custom formatter</label>
-          <input v-model="useCustomLinkFormatter" type="checkbox" />
+          <input v-model="state.useCustomLinkFormatter" type="checkbox" />
         </div>
       </div>
       <h3>v-model:</h3>
-      <div>{{ value }}</div>
+      <div>{{ state.value }}</div>
       <h3>Current Click:</h3>
-      <div>path: {{ itemPath }}</div>
+      <div>path: {{ state.itemPath }}</div>
       <div>
         data:
-        <pre>{{ itemData }}</pre>
+        <pre>{{ state.itemData }}</pre>
       </div>
     </div>
     <div class="block">
       <h3>vue-json-pretty:</h3>
       <vue-json-pretty
-        v-if="renderOK"
-        v-model="value"
-        :data="data"
-        :path="path"
-        :deep="deep"
-        :show-double-quotes="showDoubleQuotes"
-        :highlight-mouseover-node="highlightMouseoverNode"
-        :highlight-selected-node="highlightSelectedNode"
-        :show-length="showLength"
-        :show-line="showLine"
-        :select-on-click-node="selectOnClickNode"
-        :collapsed-on-click-brackets="collapsedOnClickBrackets"
-        :path-selectable="(path, data) => typeof data !== 'number'"
-        :selectable-type="selectableType"
-        :show-select-controller="showSelectController"
-        :custom-value-formatter="useCustomLinkFormatter ? customLinkFormatter : null"
-        @click="handleClick(...arguments, 'complexTree')"
+        v-if="state.renderOK"
+        v-model="state.value"
+        :data="state.data"
+        :path="state.path"
+        :deep="state.deep"
+        :show-double-quotes="state.showDoubleQuotes"
+        :highlight-mouseover-node="state.highlightMouseoverNode"
+        :highlight-selected-node="state.highlightSelectedNode"
+        :show-length="state.showLength"
+        :show-line="state.showLine"
+        :select-on-click-node="state.selectOnClickNode"
+        :collapsed-on-click-brackets="state.collapsedOnClickBrackets"
+        :path-selectable="(path, data) => typeof state.data !== 'number'"
+        :selectable-type="state.selectableType"
+        :show-select-controller="state.showSelectController"
+        :custom-value-formatter="state.useCustomLinkFormatter ? customLinkFormatter : null"
+        @click="handleClick"
         @change="handleChange"
       />
     </div>
@@ -104,6 +104,7 @@
 </template>
 
 <script>
+import { reactive, watch, nextTick } from 'vue';
 import VueJsonPretty from 'src';
 
 const defaultData = {
@@ -137,8 +138,8 @@ export default {
   components: {
     VueJsonPretty,
   },
-  data() {
-    return {
+  setup() {
+    const state = reactive({
       renderOK: true,
       val: JSON.stringify(defaultData),
       data: defaultData,
@@ -157,45 +158,58 @@ export default {
       deep: 3,
       itemData: {},
       itemPath: '',
+    });
+
+    const handleClick = (path, data) => {
+      console.log('click: ', path, data);
+      state.itemPath = path;
+      state.itemData = !data ? data + '' : data; // 处理 data = null 的情况
     };
-  },
-  watch: {
-    val(newVal) {
-      try {
-        this.data = JSON.parse(this.val);
-      } catch (err) {
-        console.log('JSON ERROR');
-      }
-    },
-    selectableType(newVal) {
-      this.renderOK = false;
-      if (newVal === 'single') {
-        this.value = 'res.error';
-      } else if (newVal === 'multiple') {
-        this.value = ['res.error', 'res.data[0].title'];
-      }
-      // 重新渲染, 因为2中情况的v-model格式不同
-      this.$nextTick(() => {
-        this.renderOK = true;
-      });
-    },
-  },
-  methods: {
-    handleClick(path, data, treeName = '') {
-      console.log('click: ', path, data, treeName);
-      this.itemPath = path;
-      this.itemData = !data ? data + '' : data; // 处理 data = null 的情况
-    },
-    handleChange(newVal, oldVal) {
+
+    const handleChange = (newVal, oldVal) => {
       console.log('newVal: ', newVal, ' oldVal: ', oldVal);
-    },
-    customLinkFormatter(data, key, path, defaultFormatted) {
+    };
+
+    const customLinkFormatter = (data, key, path, defaultFormatted) => {
       if (typeof data === 'string' && data.startsWith('http://')) {
         return `<a style="color:red;" href="${data}" target="_blank">"${data}"</a>`;
       } else {
         return defaultFormatted;
       }
-    },
+    };
+
+    watch(
+      () => state.val,
+      newVal => {
+        try {
+          state.data = JSON.parse(state.val);
+        } catch (err) {
+          console.log('JSON ERROR');
+        }
+      },
+    );
+
+    watch(
+      () => state.selectableType,
+      async newVal => {
+        state.renderOK = false;
+        if (newVal === 'single') {
+          state.value = 'res.error';
+        } else if (newVal === 'multiple') {
+          state.value = ['res.error', 'res.data[0].title'];
+        }
+        // 重新渲染, 因为2中情况的v-model格式不同
+        await nextTick();
+        state.renderOK = true;
+      },
+    );
+
+    return {
+      state,
+      customLinkFormatter,
+      handleClick,
+      handleChange,
+    };
   },
 };
 </script>
