@@ -1,5 +1,5 @@
 <template>
-  <span class="vjs-tree__brackets" @click.stop="toggleBrackets">
+  <span class="vjs-tree__brackets">
     {{ data }}
   </span>
 </template>
@@ -12,15 +12,6 @@ export default {
     data: {
       required: true,
       type: String,
-    },
-    collapsedOnClickBrackets: Boolean,
-  },
-  methods: {
-    // 切换括号展开|关闭
-    toggleBrackets(e) {
-      if (this.collapsedOnClickBrackets) {
-        this.$emit('click', e);
-      }
     },
   },
 };
